Move ElasticTestCase; add tests for commands
