+++ apps/search/es_utils.py
@@ -334,7 +334,7 @@ def es_reindex_cmd(percent=100, delete=False, models=None,
     log.info('done! (total time: %s)', format_time(delta_time))
 
 
-def es_delete_cmd(index, log=log):
+def es_delete_cmd(index, interactive=False, log=log):
     """Deletes an index"""
     try:
         indexes = [name for name, count in get_indexes()]
@@ -347,7 +347,7 @@ def es_delete_cmd(index, log=log):
         log.error('Index "%s" is not a valid index.', index)
         return
 
-    if index == READ_INDEX:
+    if index == READ_INDEX and interactive:
         ret = raw_input('"%s" is a read index. Are you sure you want '
                         'to delete it? (yes/no) ' % index)
         if ret != 'yes':
